NVIDIA: SAUCE: KVM: arm64: Allow exec fault on memory mapped cacheable in VMA
BugLink: https://bugs.launchpad.net/bugs/2119656

When the Grace Hopper/Blackwell system is setup with EGM mode in
virtualization, the system memory is partitioned into two: A Host
OS visible memory and a second EGM region that is not added to
the host OS. The EGM region is assigned to the VM as its system memory
with the QEMU VMA mapped through remap_pfn_range.

Currently KVM sets up the stage-2 mapping for memory that is not
added to the kernel with device properties. It thus does not allow
support for execution fault on such region. Since the EGM memory is
mapped through remap_pfn_range and not added to the kernel, such
memory is set without execution fault support.

This patch intends to update the KVM behaviour. It is an extension
of the proposal [1] to make KVM determine whether a region should have
NORMAL memory properties based on the VMA pgprot. The KVM behavior is
changed to set a region with support of executable fault if and only
if its VMA is mapped cacheable.

The EGM memory is NORMAL system memory that is not added to the
kernel. It is safe in terms of execution fault and is expected to
display all properties of NORMAL memory. The patch enables this
use case.

Check QEMU VMA pgprot to check if it is mapped as Normal cacheable
memory and allow exec fault.

Link:
https://lore.kernel.org/lkml/20230907181459.18145-2-ankita@nvidia.com [1]

Signed-off-by: Ankit Agrawal <ankita@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Kai-Heng Feng <kaihengf@nvidia.com>
Acked-by: Koba Ko <kobak@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>
(cherry picked from commit e38eceb9688b2c39e3034a37b9ae197c9dc36f1e https://github.com/NVIDIA/NV-Kernels/tree/24.04_linux-nvidia-adv-6.8-next)
Signed-off-by: Koba Ko <kobak@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L. Soto <csoto@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>
(backported from commit b6bd6da17f5a1287951a0e74424b843512cdce2d https://github.com/NVIDIA/NV-Kernels/tree/24.04_linux-nvidia-adv-6.11-next)
[Nirmoy: s/device/s2_force_noncacheable, s/mapping_type()/FIELD_GET(PTE_ATTRINDX_MASK, pgprot_val(page_prot))]
Signed-off-by: Nirmoy Das <nirmoyd@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Matt Ochs <mochs@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off--by: Brad Figg <bfigg@nvidia.com>